Introducte FactoryWithGetPot
This way, we can reuse the GetPot interface in the factory across
many different factory types.
